LPS-152122 Proper package
