+++ modules/dxp/apps/saml/saml-admin-rest-impl/src/main/java/com/liferay/saml/admin/rest/internal/jaxrs/exception/mapper/ConfigurationExceptionMapper.java
@@ -12,11 +12,11 @@
  *
  */
 
-package com.liferay.saml.admin.rest.internal.jaxrs.mapper;
+package com.liferay.saml.admin.rest.internal.jaxrs.exception.mapper;
 
+import com.liferay.portal.kernel.module.configuration.ConfigurationException;
 import com.liferay.portal.vulcan.jaxrs.exception.mapper.BaseExceptionMapper;
 import com.liferay.portal.vulcan.jaxrs.exception.mapper.Problem;
-import com.liferay.saml.runtime.exception.EntityIdException;
 
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
@@ -33,13 +33,15 @@ import org.osgi.service.component.annotations.Component;
 	},
 	service = ExceptionMapper.class
 )
-public class EnityIdExceptionMapper
-	extends BaseExceptionMapper<EntityIdException> {
+public class ConfigurationExceptionMapper
+	extends BaseExceptionMapper<ConfigurationException> {
 
 	@Override
-	protected Problem getProblem(EntityIdException entityIdException) {
+	protected Problem getProblem(
+		ConfigurationException configurationException) {
+
 		return new Problem(
-			Response.Status.BAD_REQUEST, entityIdException.getMessage());
+			Response.Status.BAD_REQUEST, configurationException.getMessage());
 	}
 
 	@Override
